Break up RenderTester.kt into three files.
